refactor(GuidePage): 유틸리티 함수 분리
- haversineDistanceMeters를 geoUtils.ts로 이동
- 거리 계산 함수를 독립적인 유틸리티 모듈로 분리
- 코드 재사용성 및 테스트 용이성 향상
- TypeScript 타입 안전성 유지
