+++ web/src/pages/GuidePage/index.tsx
@@ -8,6 +8,7 @@ import { useTmapSDK } from './hooks/useTmapSDK';
 import { useCurrentLocation } from './hooks/useCurrentLocation';
 import { useRouteCalculation } from './hooks/useRouteCalculation';
 import type { GuidancePoint } from './hooks/useRouteCalculation';
+import { haversineDistanceMeters } from './utils/geoUtils';
 import theme from '@/styles/theme';
 
 const GuidePage = () => {
@@ -261,20 +262,6 @@ const GuidePage = () => {
       return target;
     }
     return null;
-  };
-
-  // 거리 계산 함수 (재사용)
-  const haversineDistanceMeters = (a: LocationState, b: LocationState): number => {
-    const toRad = (deg: number) => (deg * Math.PI) / 180;
-    const R = 6371000;
-    const dLat = toRad(b.latitude - a.latitude);
-    const dLon = toRad(b.longitude - a.longitude);
-    const lat1 = toRad(a.latitude);
-    const lat2 = toRad(b.latitude);
-    const sinDLat = Math.sin(dLat / 2);
-    const sinDLon = Math.sin(dLon / 2);
-    const h = sinDLat * sinDLat + Math.cos(lat1) * Math.cos(lat2) * sinDLon * sinDLon;
-    return 2 * R * Math.asin(Math.min(1, Math.sqrt(h)));
   };
 
   // 목적지 도착 여부 확인
